+++ example/lib/plugin_example/basic_content.dart
@@ -15,15 +15,17 @@ class BasicContent extends StatelessWidget {
             _sizedContainer(
               const Image(
                 image: CachedNetworkImageProvider(
-                  'http://via.placeholder.com/350x150',
+                  'https://via.placeholder.com/350x150',
                 ),
               ),
             ),
             _sizedContainer(
               CachedNetworkImage(
                 progressIndicatorBuilder: (context, url, progress) =>
-                    CircularProgressIndicator(
+                  Center(
+                    child: CircularProgressIndicator(
                   value: progress.progress,
+                  ),
                 ),
                 imageUrl:
                     'https://images.unsplash.com/photo-1532264523420-881a47db012d?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9',
@@ -33,12 +35,12 @@ class BasicContent extends StatelessWidget {
               CachedNetworkImage(
                 placeholder: (context, url) =>
                     const CircularProgressIndicator(),
-                imageUrl: 'http://via.placeholder.com/200x150',
+                imageUrl: 'https://via.placeholder.com/200x150',
               ),
             ),
             _sizedContainer(
               CachedNetworkImage(
-                imageUrl: 'http://via.placeholder.com/300x150',
+                imageUrl: 'https://via.placeholder.com/300x150',
                 imageBuilder: (context, imageProvider) => Container(
                   decoration: BoxDecoration(
                     image: DecorationImage(
@@ -57,7 +59,7 @@ class BasicContent extends StatelessWidget {
               ),
             ),
             CachedNetworkImage(
-              imageUrl: 'http://via.placeholder.com/300x300',
+              imageUrl: 'https://via.placeholder.com/300x300',
               placeholder: (context, url) => const CircleAvatar(
                 backgroundColor: Colors.amber,
                 radius: 150,
@@ -69,7 +71,7 @@ class BasicContent extends StatelessWidget {
             ),
             _sizedContainer(
               CachedNetworkImage(
-                imageUrl: 'http://notAvalid.uri',
+                imageUrl: 'https://notAvalid.uri',
                 placeholder: (context, url) =>
                     const CircularProgressIndicator(),
                 errorWidget: (context, url, error) => const Icon(Icons.error),
@@ -85,7 +87,8 @@ class BasicContent extends StatelessWidget {
             ),
             _sizedContainer(
               CachedNetworkImage(
-                imageUrl: 'http://via.placeholder.com/350x200',
+                maxHeightDiskCache: 10,
+                imageUrl: 'https://via.placeholder.com/350x200',
                 placeholder: (context, url) =>
                     const CircularProgressIndicator(),
                 errorWidget: (context, url, error) => const Icon(Icons.error),