+++ lib/src/image_provider/_image_provider_io.dart
@@ -10,6 +10,8 @@ import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import '../../cached_network_image.dart' show ImageRenderMethodForWeb;
 import 'cached_network_image_provider.dart' as image_provider;
 
+/// IO implementation of the CachedNetworkImageProvider; the ImageProvider to
+/// load network images using a cache.
 class CachedNetworkImageProvider
     extends ImageProvider<image_provider.CachedNetworkImageProvider>
     implements image_provider.CachedNetworkImageProvider {
@@ -17,10 +19,13 @@ class CachedNetworkImageProvider
   /// When the image fails to load [errorListener] is called.
   const CachedNetworkImageProvider(
     this.url, {
+    this.maxHeight,
+    this.maxWidth,
     this.scale = 1.0,
     this.errorListener,
     this.headers,
     this.cacheManager,
+    this.cacheKey,
     //ignore: avoid_unused_constructor_parameters
     ImageRenderMethodForWeb imageRenderMethodForWeb,
   })  : assert(url != null),
@@ -33,6 +38,10 @@ class CachedNetworkImageProvider
   @override
   final String url;
 
+  /// Cache key of the image to cache
+  @override
+  final String cacheKey;
+
   /// Scale of the image
   @override
   final double scale;
@@ -44,6 +53,12 @@ class CachedNetworkImageProvider
   /// Set headers for the image provider, for example for authentication
   @override
   final Map<String, String> headers;
+
+  @override
+  final int maxHeight;
+
+  @override
+  final int maxWidth;
 
   @override
   Future<CachedNetworkImageProvider> obtainKey(
@@ -76,9 +91,24 @@ class CachedNetworkImageProvider
   ) async* {
     assert(key == this);
     try {
-      var mngr = S3CacheManager();
-      await for (var result in mngr.getS3FileStream(key.url,
-          withProgress: true, headers: headers)) {
+      var mngr = cacheManager ?? DefaultCacheManager();
+      assert(
+          mngr is ImageCacheManager || (maxWidth == null && maxHeight == null),
+          'To resize the image with a CacheManager the '
+          'CacheManager needs to be an ImageCacheManager. maxWidth and '
+          'maxHeight will be ignored when a normal CacheManager is used.');
+
+      var stream = mngr is ImageCacheManager
+          ? mngr.getImageFile(key.url,
+              maxHeight: maxHeight,
+              maxWidth: maxWidth,
+              withProgress: true,
+              headers: headers,
+              key: key.cacheKey)
+          : mngr.getFileStream(key.url,
+              withProgress: true, headers: headers, key: key.cacheKey);
+
+      await for (var result in stream) {
         if (result is DownloadProgress) {
           chunkEvents.add(ImageChunkEvent(
             cumulativeBytesLoaded: result.downloaded,
@@ -110,13 +140,16 @@ class CachedNetworkImageProvider
   @override
   bool operator ==(dynamic other) {
     if (other is CachedNetworkImageProvider) {
-      return url == other.url && scale == other.scale;
+      return ((cacheKey ?? url) == (other.cacheKey ?? other.url)) &&
+          scale == other.scale &&
+          maxHeight == other.maxHeight &&
+          maxWidth == other.maxWidth;
     }
     return false;
   }
 
   @override
-  int get hashCode => hashValues(url, scale);
+  int get hashCode => hashValues(cacheKey ?? url, scale, maxHeight, maxWidth);
 
   @override
   String toString() => '$runtimeType("$url", scale: $scale)';