+++ lib/src/image_provider/cached_network_image_provider.dart
@@ -1,13 +1,16 @@
 import 'package:cached_network_image/src/cache_manager/s3_cache_manager.dart';
 import 'package:flutter/painting.dart';
 import 'package:flutter/widgets.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
+
 import '_image_provider_io.dart'
     if (dart.library.html) '_image_provider_web.dart' as image_provider;
 
+/// Function which is called after loading the image failed.
 typedef ErrorListener = void Function();
 
 /// Currently there are 2 different ways to show an image on the web with both
-/// their own pros and cons, using a custom [HttpGet] (the default for this library)
+/// their own pros and cons, using a custom [HttpGet]
 /// or an HTML Image element mentioned [here on a GitHub issue](https://github.com/flutter/flutter/issues/57187#issuecomment-635637494).
 ///
 /// When using HttpGet the image will work on Skia and it will use the [CachedNetworkImageProvider.headers]
@@ -18,10 +21,16 @@ typedef ErrorListener = void Function();
 /// The [HtmlImage] does not need a CORS handshake, but it also does not use your
 /// provided headers and it does not work when using Skia to render the page.
 enum ImageRenderMethodForWeb {
+  /// HtmlImage uses a default web image including default browser caching.
+  /// This is the recommended and default choice.
   HtmlImage,
+
+  /// HttpGet uses an http client to fetch an image. It enables the use of
+  /// headers, but loses some default web functionality.
   HttpGet,
 }
 
+/// An ImageProvider to load images from the network with caching functionality.
 abstract class CachedNetworkImageProvider
     extends ImageProvider<CachedNetworkImageProvider> {
   /// Creates an object that fetches the image at the given URL.
@@ -32,6 +41,9 @@ abstract class CachedNetworkImageProvider
   /// for the benefits of each method.
   const factory CachedNetworkImageProvider(
     String url, {
+    int maxHeight,
+    int maxWidth,
+    String cacheKey,
     double scale,
     @Deprecated('ErrorListener is deprecated, use listeners on the imagestream')
         ErrorListener errorListener,
@@ -46,17 +58,30 @@ abstract class CachedNetworkImageProvider
   /// When running flutter on the web, the cacheManager is not used.
   S3CacheManager get cacheManager;
 
+  /// The errorListener is called when the ImageProvider failed loading the
+  /// image. Deprecated in favor of [ImageStreamListener.onError].
   @deprecated
   ErrorListener get errorListener;
 
   /// The URL from which the image will be fetched.
   String get url;
 
+  /// The Key from image for cache
+  String get cacheKey;
+
   /// The scale to place in the [ImageInfo] object of the image.
   double get scale;
 
   /// The HTTP headers that will be used to fetch image from network.
   Map<String, String> get headers;
+
+  /// Max height in pixels for the image. When set the resized image is
+  /// stored in the cache.
+  int get maxHeight;
+
+  /// Max width in pixels for the image. When set the resized image is
+  /// stored in the cache.
+  int get maxWidth;
 
   @override
   ImageStreamCompleter load(